requirements.txt
# Core FastAPI
fastapi==0.116.1
uvicorn[standard]==0.23.2

# ORM & Database
sqlalchemy==2.0.43
pymysql==1.1.1
alembic==1.16.4

# Environment variables
python-dotenv==1.0.0
python-multipart-0.0.20

# Data validation (Pydantic v1)
pydantic<2.0

# Authentication & Security
passlib[bcrypt]==1.7.5
python-jose==3.3.0

# Email validation
email-validator==2.2.0
